Move shared connectivity logic to shared library
